Up the seed size to uint64_t
These is probably overkill, but it's easier to do this now than later.

+++ crawl-ref/source/options.h
@@ -170,7 +170,7 @@ public:
                                 // should be accessible by different people.
     vector<string> additional_macro_files;
 
-    uint32_t    seed;   // Non-random games.
+    uint64_t    seed;   // Non-random games.
 
 #ifdef DGL_SIMPLE_MESSAGING
     bool        messaging;      // Check for messages.
